[DataLayer] Prepare beta version of the Adobe Client Data Layer
- moves jest config under gulp/configs
- removes unnecessary use-strict in module
