+++ gulp/configs/paths.config.js
