+++ gulp/tasks/test.js
@@ -10,13 +10,14 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 module.exports = function(gulp) {
-    'use strict';
-
     const jestcli = require('jest-cli');
-    const config = require(`${__dirname}/../../jest.config.js`);
+
+    const config = {
+      jest: require(`${__dirname}/../configs/jest.config.js`)
+    };
 
     gulp.task('test', (done) => {
-        jestcli.runCLI(config,['.']);
+        jestcli.runCLI(config.jest,['.']);
         done();
     });
 };
