--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-    collectCoverage: true,
-    coverageDirectory: '<rootDir>/test/unit-test-coverage',
-    testMatch: ['<rootDir>/src/tests/**/*Test.js']
-};
